River-71 Patch, Thankyou Sim.
Bugtraq ID 6449005

An RFE for Reggie: make it so that the ServerSocket used to
handle unicast discovery is created from a configurable ServerSocketFactory.
We've run across this when trying to run Reggie behind a firewall where a range
of ports is exposed--if there were a configurable ServerSocketFactory, then we
could specify one that allocated a ServerSocket within that range.

(I realize we could configure a fixed port number or open up the default
unicast discovery port in the firewall, but we'd like to avoid hardcoding
portnumbers into our configurations as much as possible, and also don't want to
preclude running more than one instance of Reggie on a given host.)

git-svn-id: https://svn.apache.org/repos/asf/incubator/river/jtsk/trunk@881217 13f79535-47bb-0310-9956-ffa450edef68