This patch integrates changes to replace the RMIClassLoaderSPI with a CodebaseAccessClassLoader as per River-336. Although it doesn't include all changes to PreferredClassProvider.
Additional changes to net.jini.io have been made to allow a subset of this package to be built at a later stage for CDC Personal Profile 1.11 (a subset of Java 1.4.2).

An additional interface has been created, which is a subset of ServiceRegistrar, which ServiceRegistrar now extends, this is to enable building a release for CDC Personal Profile 1.11 or greater at a later point in time, without causing any breakage to existing application software.

Furthermore additional permissions have been granted to the qa test defaulttest.policy in order for some random test failures to pass.  This was specific to the Java 1.6 platform and I believe related to the -source=5 and -target=jsr14 compiler optimisations that load a security manager earlier, it is also related to the removal of the need to utilise the jre/etc directory for jsk-policy.jar

These changes are experimental, and some incomplete, please feel free to make comments and reccommendations or hightlite any unforseen issues.

The Apache River version has been changed to 2.2.0


git-svn-id: https://svn.apache.org/repos/asf/incubator/river/jtsk/trunk@934258 13f79535-47bb-0310-9956-ffa450edef68