River Refactoring, some minor refactoring to ConcurrentDynamicPolicyProvider.
Renamed DiscoveryManagement2 interface to RegistrarManagement, which supercedes DiscoveryManagement.

Removed the Facade's for DiscoveryManagement, the Facade was undesireable since, Discovery Managers usually implement a number of interfaces.

These changes are to enable wider Java platform support by making minor changes to the Jini API by deprecating methods and interfaces that are java.rmi specific, this will later enable a subset of the Jini Platform to be supported on Java CDC Personal Basis Profile, to allow participation in a djinn using services.


git-svn-id: https://svn.apache.org/repos/asf/incubator/river/jtsk/trunk@962760 13f79535-47bb-0310-9956-ffa450edef68