River-345 patch merged into ServiceDiscoverManager
Serialization Issue fixed in RemoteEvent and RemoteDiscoveryEvent, unit tests included.

Null pointer fixed in RemoteEvent, when null MarhalledObject is passed in and converted to a null MarshalledInstance.

Other changes relate to new method signatures, and the ExecutionContextManager.

The ExecutionContextManager will probably be removed.  The RevokeableDynamicPolicy will be moved out of the Public API, since this is still experimental.

git-svn-id: https://svn.apache.org/repos/asf/incubator/river/jtsk/trunk@990669 13f79535-47bb-0310-9956-ffa450edef68