River-396 PreferredClassProvider Concurrency improvement.  Re-factored and simplified the weak cache, which was a little difficult to understand due to the relations between weak references and WeakHashMap.
WeakHashMap uses weak keys but what we really wanted was a ConcurrentMap that used weak values, where the value is a child ClassLoader, the key is unique to the ClassLoader so when it becomes weakly reachable the key can be removed too.

using a ConcurrentMap also eliminates any issue with synchronisation and atomic replacement of a key & value combination that has become weakly reachable.  This allows removal of the synchronised block around checking ClassLoader permissions, Concurrency in practise recommends against synchronising on external state.

Tests have been revised to include the classpath and permissions for the new libraries.

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/trunk@1366641 13f79535-47bb-0310-9956-ffa450edef68