Fix index out of range for LookupServices
Changed waitForDiscovery() method to wait on object monitor for discovery events with a timeout.  I suspect that the synchronized block inside the loop and following loop can optimised outside the loop by the compiler, waiting on the object monitor ensures the lock is released, so that discovery events can be received.

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/skunk/qa_refactor/trunk@1439020 13f79535-47bb-0310-9956-ffa450edef68