Updated version numbers and @since version annotations in preparation for new release. Set skip test verifiers on failing tests prior to release, so we can identify any new test failures during development while waiting for fixes for existing test failures.
Fixed issue related to DNS lookups caused by SecureClassLoader using CodeSource as a key in an internal ProtectionDomain cache, relates to River-401.
http://bugs.sun.com/view_bug.do?bug_id=6826789
http://bugs.sun.com/view_bug.do?bug_id=7152564

Refactored PermissionGrant implementation and PermissionGrantBuilder api to use a String for uri, rather than the URI class, this allows other implementations to be used.

Replaced UriString with new Uri class in UriGrant, which has resulted in a 5x localised performance increase.

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/skunk/qa_refactor/trunk@1463106 13f79535-47bb-0310-9956-ffa450edef68