Cleaned up synchronization and locking strategies and removed instances of synchronization while performing remote calls.
Changed final fields to non final where "this" escaped during construction.

Replaced TaskManager with ExecutorService where Task.runAfter was not required, TaskManager is still utilised by JoinManager, ServiceDiscoveryManager and the qa test framework.
Changed TaskManager to accept Runnable and to check for Task instances.

Changed RetryTask, pushed TaskManager.Task implementations to subclasses that used it.  RetryTask is not compatible with Task.runAfter as documented in River-344.

Added Generics to static configuration methods, to eliminate casts.

Started working on ScheduledExecutors to take advantage of a ScheduledExecutorService to allow sharing of a thread pool for tasks that have dependencies, require order and must not execute until their dependencies have completed.

git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/skunk/qa_refactor/trunk@1554723 13f79535-47bb-0310-9956-ffa450edef68