Skip MultihomedClientTest - seems to be having problems with sun name service provider spi. This is scheduled for removal in Java 9, however some functionality has been provided as per below.
Reference material:

https://bugs.openjdk.java.net/browse/JDK-8134577

To facilitate testing activity, and provide a replacement for the
customized NameService implementations in the
JDK tests, the default NameService has been extended to support
the retrieval of host to IP address mappings from a file.
The file path is specified with a system property " jdk.internal.hosts".

Previously a nameservice provider was specified by setting the system
property
"sun.net.spi.nameservice.provider.", as per the documentation
http://docs.oracle.com/javase/8/docs/technotes/guides/net/properties.html


git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/trunk@1719767 13f79535-47bb-0310-9956-ffa450edef68