Initial commit, note that there is work in here that was intended to be in the River 3.0.0 release, with improvements to documentation and some synchronization visibility correctness updates for RegistrarImpl, however I need to check atomicity is correct.  These are not related to the security changes, but were lumped into the same patch at the time when this work was abandoned due to difficulties on river dev.
@AtomicSerial, note that the trusted classes hasn't been completely implemented yet.  AtomicSerial uses a custom ObjectInputStream to validate input, trusted classes will limit the classes that can be deserialized, for those with conventional Serialization that don't implement @AtomicSerial, but have been deemed safe.

DeSerializationPermission, will limit the classes that can be deserialized.

The intent is that a new default lookup method in ServiceRegistrar will return bootstrap proxy's, based on service matches, as well as services with compliant constraints.

A bootstrap proxy is local code, that has DeSerializationPermission, during proxy preparation, the bootstrap proxy will be asked for the service proxy, after authentication using a bootstrap proxy, the codebase for the service proxy will be dynamically granted DownloadPermission and DeSerializationPermission.