net.jini.lookup.LookupCacheImpl serviceIdMap access not atomic during lookup call.  #8
Separate ServiceItemReg from LookupCacheImpl for re-implementation.  Make all access to serviceIdMap obvious for easier review and future refactoring if necessary.  Reactivate section of code in lookup method, that removes ServiceItemReg's from serviceIdMap, utilise BiFunction to ensure ServiceReg addition, mutation and discard operations are atomic.

Possible future immutable implementation of ServiceItemReg => ServiceReg.

Change EventReg additions to eventRegMap to be made after lookup to avoid concurrent modification with events. Also modify ServiceEvent handling so that events aren't processed concurrently with lookup local state synchronization with the ServiceRegistrar.  Introduced a new ExecutorService to handle ServiceEvent processing.  Reduced the instances where lookup is used too readily, due to the cost of a lookup call, instead passing non contiguous ServiceEvent's back to the queue, allowing time for earlier events to arrive in the queue, a timeout of 10 seconds is used, after which a lookup call will be made to synchronized local state with the remote ServiceRegistrar's.   ServiceEvent's that arrive too early are passed back to the ExecutorService for later processing.