net.jini.lookup.LookupCacheImpl serviceIdMap access not atomic during lookup call.  #8
Change EventReg additions to eventRegMap to be made after lookup to avoid concurrent modification with events. - Was part of previous commit, the current changes were made after reviewing removeProxyReg and DropProxyRegTask.  This is to ensure that, DropProxyRegTask, RegisterListenerTask, lookup and ServiceEvent processing all are mutually exclusive processes, this ensures the state of serviceIdMap doesn't become corrupted and remains a correct filtered mirror of the ServiceRegistrar.