Rearrange modules to avoid split packages, move some non API classes to new packages to avoid split packages between modules.
Changed name to JGDMS to avoid trademark issues with Apache River, at least for now until this code is donated back to Apache River.

Proxy and service implementations are now in separate packages to avoid split packages, this means that some proxy implementation classes are now public, this is not seen as an issue however, as these classes were Serializable and hence already effectively published publicly by Serialization.