Discovery V2, Unicast discovery fails to read input bytebuffer fully, resulting in hash mismatch for secure discovery providers #68
Previous fix was untested, and failed on testing, although it appears very similar, it seems it's important that the last UnicastDiscoveryServer that satisfies constraints is selected, as well as fully reading the bytebuffer.   The bytebuffer must be fully read in to ensure the hash matches.

This is the original tested fix.